boost::ublas: gcc support for optimal matrix multiplication
This patch contains an optimised, vectorised kernel, using gcc's
SIMD vectors.  This reaches matrix multiplication speeds
comparable to hand-crafted assembly kernels on x86 Haswell
microarchitecture.

The kernels are compile-time parametrisable, ans the patch also
contains optimised kernel parameters for float, double, and long
double.

The validity of the parameters are checked during compile-time,
but the produced compile error can be relatively obscure if
vector_length is too big.

For architectures that doesn't support vectorisation for all types, it is
possible to specify parameters that cause really suboptimal code generation.
But as far as I see code correctness is ensured.

Signed-off-by: Imre Palik <imre_palik@yahoo.co.uk>
Cc: Michael Lehn <michael.lehn@uni-ulm.de>